+++ nuclia/sdk/search.py
@@ -1,3 +1,4 @@
+import backoff
 import json
 import os
 import warnings
@@ -18,6 +19,7 @@ from nucliadb_models.search import (
     SearchRequest,
     SyncAskResponse,
 )
+from nucliadb_sdk import exceptions
 from pydantic import ValidationError
 
 from nuclia.data import get_async_auth, get_auth
@@ -65,6 +67,13 @@ class NucliaSearch:
 
     @kb
     @pretty
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def search(
         self,
         *,
@@ -95,6 +104,13 @@ class NucliaSearch:
 
     @kb
     @pretty
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def find(
         self,
         *,
@@ -134,6 +150,13 @@ class NucliaSearch:
         return ndb.ndb.find(req, kbid=ndb.kbid)
 
     @kb
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def ask(
         self,
         *,
@@ -198,6 +221,13 @@ class NucliaSearch:
         return result
 
     @kb
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def ask_json(
         self,
         schema: Union[str, Dict[str, Any]],